- moved `displayModal` calculator window from `VulnerabilityHeader` to `VulnerabilityBadge` - fixed that calculator wasn't in center of screen - fixed unnecessary multiple checking currentUser`s rights for editing vulnerability
