+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/vuln/CreateVulnerabilityView.kt
@@ -16,8 +16,8 @@ import com.saveourtool.save.frontend.components.basic.renderAvatar
 import com.saveourtool.save.frontend.components.inputform.InputTypes
 import com.saveourtool.save.frontend.components.inputform.inputTextFormOptional
 import com.saveourtool.save.frontend.components.modal.MAX_Z_INDEX
+import com.saveourtool.save.frontend.components.modal.calculatorModalStyle
 import com.saveourtool.save.frontend.components.modal.displayModal
-import com.saveourtool.save.frontend.components.modal.largeTransparentModalStyle
 import com.saveourtool.save.frontend.components.views.vuln.component.cvssBaseScoreCalculator
 import com.saveourtool.save.frontend.components.views.vuln.component.uploadCosvButton
 import com.saveourtool.save.frontend.externals.fontawesome.*
@@ -146,7 +146,7 @@ val createVulnerabilityView: VFC = VFC {
                 }
             }
         },
-        modalStyle = largeTransparentModalStyle,
+        modalStyle = calculatorModalStyle,
         onCloseButtonPressed = cvssCalculatorWindowOpenness.closeWindowAction(),
         customWidth = jso { width = 45.rem }
     ) { }
