+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/vuln/VulnerabilityBadge.kt
@@ -2,14 +2,25 @@
 
 package com.saveourtool.save.frontend.components.views.vuln
 
+import com.saveourtool.save.cvsscalculator.CvssVersion
+import com.saveourtool.save.cvsscalculator.parsingVectorToMap
+import com.saveourtool.save.cvsscalculator.v3.CvssVectorV3
 import com.saveourtool.save.entities.cosv.VulnerabilityExt
+import com.saveourtool.save.frontend.components.modal.calculatorModalStyle
+import com.saveourtool.save.frontend.components.modal.displayModal
+import com.saveourtool.save.frontend.components.views.vuln.component.cvssBaseScoreCalculator
 import com.saveourtool.save.frontend.externals.i18next.useTranslation
 import com.saveourtool.save.frontend.externals.progressbar.progressBar
 import com.saveourtool.save.frontend.themes.Colors
 import com.saveourtool.save.frontend.utils.buttonBuilder
+import com.saveourtool.save.frontend.utils.useWindowOpenness
+
+import com.saveourtool.osv4k.Severity
+import com.saveourtool.osv4k.SeverityType
 import js.core.jso
 import react.FC
 import react.Props
+import react.StateSetter
 import react.dom.html.ReactHTML.a
 import react.dom.html.ReactHTML.div
 import react.dom.html.ReactHTML.h4
@@ -24,6 +35,66 @@ private const val MAX_VALUE = 10.0f
 val vulnerabilityBadge: FC<VulnerabilityBadgeProps> = FC { props ->
     val (t) = useTranslation("vulnerability")
     val severityNum = props.vulnerability.metadataDto.severityNum
+    val cvssCalculatorWindowOpenness = useWindowOpenness()
+
+    @Suppress("EMPTY_BLOCK_STRUCTURE_ERROR")
+    displayModal(
+        cvssCalculatorWindowOpenness.isOpen(),
+        "Base Score Calculator".t(),
+        bodyBuilder = {
+            cvssBaseScoreCalculator {
+                baseMetricsOfVulnerability = props.vulnerability.cosv.severity?.let { severityList ->
+                    val severityV3Count = severityList.count { it.type == SeverityType.CVSS_V3 }
+
+                    if (severityV3Count == 1) {
+                        val severityV3 = severityList.find { it.type == SeverityType.CVSS_V3 }!!
+                        CvssVectorV3(severityV3.score.parsingVectorToMap()).baseMetrics
+                    } else {
+                        undefined
+                    }
+                } ?: undefined
+
+                onCloseButton = { cvssCalculatorWindowOpenness.closeWindow() }
+
+                onCloseButtonPassed = { baseMetrics ->
+                    props.vulnerability.cosv.severity?.let { currentSeverityList ->
+                        val severityV2Count = currentSeverityList.count { it.type == SeverityType.CVSS_V2 }
+                        val severityList = currentSeverityList.toMutableList()
+
+                        if (severityV2Count > 1) {
+                            severityList.removeAll { it.type == SeverityType.CVSS_V2 }
+                        }
+
+                        val cvssVector = CvssVectorV3(CvssVersion.CVSS_V3_1, baseMetrics)
+                        severityList.removeAll { it.type == SeverityType.CVSS_V3 }
+                        severityList.add(
+                            Severity(
+                                type = SeverityType.CVSS_V3,
+                                score = cvssVector.scoreVectorString(),
+                                scoreNum = cvssVector.calculateBaseScore().toString(),
+                            )
+                        )
+
+                        props.setVulnerability { it?.copy(cosv = props.vulnerability.cosv.copy(severity = severityList.toList())) }
+                        props.setVulnerability { it?.copy(metadataDto = props.vulnerability.metadataDto.copy(severityNum = cvssVector.calculateBaseScore())) }
+                    } ?: run {
+                        val cvssVector = CvssVectorV3(CvssVersion.CVSS_V3_1, baseMetrics)
+                        val severityList = listOf(Severity(
+                            type = SeverityType.CVSS_V3,
+                            score = cvssVector.scoreVectorString(),
+                            scoreNum = cvssVector.calculateBaseScore().toString(),
+                        ))
+
+                        props.setVulnerability { it?.copy(cosv = props.vulnerability.cosv.copy(severity = severityList)) }
+                        props.setVulnerability { it?.copy(metadataDto = props.vulnerability.metadataDto.copy(severityNum = cvssVector.calculateBaseScore())) }
+                    }
+                }
+            }
+        },
+        modalStyle = calculatorModalStyle,
+        onCloseButtonPressed = cvssCalculatorWindowOpenness.closeWindowAction(),
+        customWidth = jso { width = 45.rem }
+    ) { }
 
     val (color, criticalityLabel) = when (severityNum) {
         in 0f..FOR_GREEN -> Colors.SUCCESS.value to "Low"
@@ -87,9 +158,9 @@ val vulnerabilityBadge: FC<VulnerabilityBadgeProps> = FC { props ->
                     className = ClassName("col-2 ml-4")
                     div {
                         className = ClassName("row align-items-start justify-content-end")
-                        props.onEditClick?.let { onClickCallback ->
+                        if (!props.isEditDisabled) {
                             buttonBuilder("Edit".t(), style = "primary", isOutline = false) {
-                                onClickCallback()
+                                cvssCalculatorWindowOpenness.openWindow()
                             }
                         }
                     }
@@ -109,7 +180,12 @@ external interface VulnerabilityBadgeProps : Props {
     var vulnerability: VulnerabilityExt
 
     /**
-     * Callback that should be invoked on edit button click
+     * Vulnerability setter
      */
-    var onEditClick: (() -> Unit)?
+    var setVulnerability: StateSetter<VulnerabilityExt?>
+
+    /**
+     * If the edit button was not yet pressed
+     */
+    var isEditDisabled: Boolean
 }
