+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/vuln/VulnerabilityGeneralInfoProps.kt
@@ -54,9 +54,8 @@ val vulnerabilityGeneralInfoProps: FC<VulnerabilityGeneralInfoProps> = FC { prop
                         }
                     }
                     if (props.isEditDisabled) {
-                        // only Super Users and owners of unapproved vuln. can edit it
-                        if (props.userInfo?.isSuperAdmin() == true ||
-                                (props.userInfo?.name == metadataDto.user.name && props.vulnerability.metadataDto.status != VulnerabilityStatus.APPROVED)) {
+                        // only Super Users or owners/participants of unapproved vulnerability can edit it
+                        if (hasRightsToEdit(props.currentUserInfo, this@with)) {
                             buttonBuilder(
                                 labelBuilder = {
                                     p {
@@ -190,7 +189,7 @@ val vulnerabilityGeneralInfoProps: FC<VulnerabilityGeneralInfoProps> = FC { prop
 
                 // ================= tags ===================
 
-                if (props.canEditVulnerability || props.vulnerability.metadataDto.tags.isNotEmpty()) {
+                if (!props.isEditDisabled || props.vulnerability.metadataDto.tags.isNotEmpty()) {
                     hr { }
                     h6 {
                         className = ClassName("font-weight-bold text-primary-blue mb-4")
@@ -198,9 +197,8 @@ val vulnerabilityGeneralInfoProps: FC<VulnerabilityGeneralInfoProps> = FC { prop
                     }
                     div {
                         vulnerabilityTagsComponent {
-                            this.userInfo = props.userInfo
+                            this.currentUserInfo = props.currentUserInfo
                             this.vulnerability = props.vulnerability
-                            this.canEditVulnerability = props.canEditVulnerability
                             this.isEditDisabled = props.isEditDisabled
                             this.setVulnerabilityExt = props.setVulnerability
                         }
@@ -319,14 +317,9 @@ external interface VulnerabilityGeneralInfoProps : Props {
     var fetchVulnerability: () -> Unit
 
     /**
-     * FLag that defines if current user can edit vulnerability or not
-     */
-    var canEditVulnerability: Boolean
-
-    /**
      * Currently logged-in user or null
      */
-    var userInfo: UserInfo?
+    var currentUserInfo: UserInfo?
 
     /**
      * If the edit button was not yet pressed
@@ -348,3 +341,12 @@ external interface VulnerabilityGeneralInfoProps : Props {
      */
     var setSelectedMenu: StateSetter<VulnerabilityTab>
 }
+
+/**
+ * @param currentUserInfo
+ * @param vulnerability
+ * @return if user can press edit button
+ */
+private fun hasRightsToEdit(currentUserInfo: UserInfo?, vulnerability: VulnerabilityExt): Boolean =
+        currentUserInfo?.isSuperAdmin() == true || ((currentUserInfo?.name == vulnerability.metadataDto.user.name ||
+                currentUserInfo?.name in vulnerability.getAllParticipants().map { it.name }) && vulnerability.metadataDto.status != VulnerabilityStatus.APPROVED)
