+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/vuln/VulnerabilityHeader.kt
@@ -6,18 +6,13 @@
 
 package com.saveourtool.save.frontend.components.views.vuln
 
-import com.saveourtool.save.cvsscalculator.CvssVersion
-import com.saveourtool.save.cvsscalculator.parsingVectorToMap
-import com.saveourtool.save.cvsscalculator.v3.*
 import com.saveourtool.save.entities.CommentDto
 import com.saveourtool.save.entities.cosv.VulnerabilityExt
 import com.saveourtool.save.entities.vulnerability.VulnerabilityStatus
 import com.saveourtool.save.frontend.components.inputform.InputTypes
 import com.saveourtool.save.frontend.components.modal.displayModal
-import com.saveourtool.save.frontend.components.modal.largeTransparentModalStyle
 import com.saveourtool.save.frontend.components.modal.mediumTransparentModalStyle
 import com.saveourtool.save.frontend.components.views.contests.tab
-import com.saveourtool.save.frontend.components.views.vuln.component.cvssBaseScoreCalculator
 import com.saveourtool.save.frontend.externals.fontawesome.faDownload
 import com.saveourtool.save.frontend.externals.fontawesome.faImage
 import com.saveourtool.save.frontend.externals.fontawesome.faTable
@@ -27,8 +22,6 @@ import com.saveourtool.save.frontend.utils.*
 import com.saveourtool.save.info.UserInfo
 import com.saveourtool.save.validation.FrontendRoutes
 
-import com.saveourtool.osv4k.Severity
-import com.saveourtool.osv4k.SeverityType
 import js.core.jso
 import react.*
 import react.dom.aria.ariaDescribedBy
@@ -50,7 +43,6 @@ internal val headerMenu: FC<HeaderMenuProps> = FC { props ->
     val (comment, setComment) = useState(CommentDto.empty)
     val (buttonType, setButtonType) = useState<ButtonType?>(null)
     val deleteVulnerabilityWindowOpenness = useWindowOpenness()
-    val cvssCalculatorWindowOpenness = useWindowOpenness()
 
     // ======================= requests ================================================================================
 
@@ -197,73 +189,14 @@ internal val headerMenu: FC<HeaderMenuProps> = FC { props ->
         }
     }
 
-    @Suppress("EMPTY_BLOCK_STRUCTURE_ERROR")
-    displayModal(
-        cvssCalculatorWindowOpenness.isOpen(),
-        "Base Score Calculator".t(),
-        bodyBuilder = {
-            cvssBaseScoreCalculator {
-                baseMetricsOfVulnerability = props.vulnerability.cosv.severity?.let { severityList ->
-                    val severityV3Count = severityList.count { it.type == SeverityType.CVSS_V3 }
-
-                    if (severityV3Count == 1) {
-                        val severityV3 = severityList.find { it.type == SeverityType.CVSS_V3 }!!
-                        CvssVectorV3(severityV3.score.parsingVectorToMap()).baseMetrics
-                    } else {
-                        undefined
-                    }
-                } ?: undefined
-
-                onCloseButton = { cvssCalculatorWindowOpenness.closeWindow() }
-
-                onCloseButtonPassed = { baseMetrics ->
-                    props.vulnerability.cosv.severity?.let { currentSeverityList ->
-                        val severityV2Count = currentSeverityList.count { it.type == SeverityType.CVSS_V2 }
-                        val severityList = currentSeverityList.toMutableList()
-
-                        if (severityV2Count > 1) {
-                            severityList.removeAll { it.type == SeverityType.CVSS_V2 }
-                        }
-
-                        val cvssVector = CvssVectorV3(CvssVersion.CVSS_V3_1, baseMetrics)
-                        severityList.removeAll { it.type == SeverityType.CVSS_V3 }
-                        severityList.add(
-                            Severity(
-                                type = SeverityType.CVSS_V3,
-                                score = cvssVector.scoreVectorString(),
-                                scoreNum = cvssVector.calculateBaseScore().toString(),
-                            )
-                        )
-
-                        props.setVulnerability { it?.copy(cosv = props.vulnerability.cosv.copy(severity = severityList.toList())) }
-                        props.setVulnerability { it?.copy(metadataDto = props.vulnerability.metadataDto.copy(severityNum = cvssVector.calculateBaseScore())) }
-                    } ?: run {
-                        val cvssVector = CvssVectorV3(CvssVersion.CVSS_V3_1, baseMetrics)
-                        val severityList = listOf(Severity(
-                            type = SeverityType.CVSS_V3,
-                            score = cvssVector.scoreVectorString(),
-                            scoreNum = cvssVector.calculateBaseScore().toString(),
-                        ))
-
-                        props.setVulnerability { it?.copy(cosv = props.vulnerability.cosv.copy(severity = severityList)) }
-                        props.setVulnerability { it?.copy(metadataDto = props.vulnerability.metadataDto.copy(severityNum = cvssVector.calculateBaseScore())) }
-                    }
-                }
-            }
-        },
-        modalStyle = largeTransparentModalStyle,
-        onCloseButtonPressed = cvssCalculatorWindowOpenness.closeWindowAction(),
-        customWidth = jso { width = 45.rem }
-    ) { }
-
     // ======================= rendering ===============================================================================
 
     div {
         className = ClassName("col-3 mr-3")
         vulnerabilityBadge {
             this.vulnerability = props.vulnerability
-            this.onEditClick = { cvssCalculatorWindowOpenness.openWindow() }
-                .takeIf { !props.isEditDisabled }
+            this.setVulnerability = props.setVulnerability
+            this.isEditDisabled = props.isEditDisabled
         }
     }
     div {
