+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/vuln/VulnerabilityInfoTab.kt
@@ -49,10 +49,6 @@ val vulnerabilityInfoTab: FC<VulnerabilityInfoTabProps> = FC { props ->
     val deleteVulnerabilityWindowOpenness = useWindowOpenness()
 
     val (deleteProject, setDeleteProject) = useState<CosvAffected?>(null)
-
-    // user can be able to edit the view only if he has rights and has pressed edit button
-    val isAbleToEdit = hasRightsToEdit(props.currentUserInfo, props.vulnerability) && !props.isEditDisabled
-
     val (vulnerabilityProjects, setVulnerabilityProjects) = useStateFromProps(props.vulnerability.cosv.affected ?: emptyList())  // <List<VulnerabilityProjectDto>>(emptyList())
 
     val enrollRequest = useDeferredRequest {
@@ -256,7 +252,7 @@ val vulnerabilityInfoTab: FC<VulnerabilityInfoTabProps> = FC { props ->
             )
 
         onAddClick = { dateWindowOpenness.openWindow() }
-            .takeIf { isAbleToEdit }
+            .takeIf { !props.isEditDisabled }
 
         onNodeClick = { date: LocalDateTime, type: String ->
             props.vulnerability.cosv.getTimeline()
@@ -303,7 +299,7 @@ val vulnerabilityInfoTab: FC<VulnerabilityInfoTabProps> = FC { props ->
                 }
                 ?: Unit
         }
-            .takeIf { isAbleToEdit }
+            .takeIf { !props.isEditDisabled }
     }
 
     @Suppress("COMPLEX_EXPRESSION")
@@ -313,7 +309,7 @@ val vulnerabilityInfoTab: FC<VulnerabilityInfoTabProps> = FC { props ->
         vulnerabilityProjects,
         projectTable,
         props.isTableView,
-        isAbleToEdit,
+        !props.isEditDisabled,
         t,
     ) {
         setRequestedType(it)
@@ -326,7 +322,7 @@ val vulnerabilityInfoTab: FC<VulnerabilityInfoTabProps> = FC { props ->
         vulnerabilityProjects,
         commitFixTable,
         props.isTableView,
-        isAbleToEdit,
+        !props.isEditDisabled,
         t,
     ) {
         setRequestedType(it)
