+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/vuln/VulnerabilityTagsComponent.kt
@@ -39,7 +39,7 @@ val vulnerabilityTagsComponent: FC<VulnerabilityTagsComponentProps> = FC { props
         )
     }
 
-    if (props.canEditVulnerability && !props.isEditDisabled) {
+    if (!props.isEditDisabled) {
         div {
             className = ClassName("input-group shadow mb-1")
             input {
@@ -69,7 +69,7 @@ val vulnerabilityTagsComponent: FC<VulnerabilityTagsComponentProps> = FC { props
     div {
         className = ClassName("col pl-0")
         props.vulnerability.metadataDto.tags.forEach { tag ->
-            val tagEditable = props.userInfo?.name == props.vulnerability.metadataDto.user.name
+            val tagEditable = props.currentUserInfo?.name == props.vulnerability.metadataDto.user.name
             buttonBuilder(
                 tag,
                 isOutline = true,
@@ -99,11 +99,6 @@ val vulnerabilityTagsComponent: FC<VulnerabilityTagsComponentProps> = FC { props
  */
 external interface VulnerabilityTagsComponentProps : Props {
     /**
-     * FLag that defines if current user can edit vulnerability or not
-     */
-    var canEditVulnerability: Boolean
-
-    /**
      * Current vulnerability
      */
     var vulnerability: VulnerabilityExt
@@ -116,7 +111,7 @@ external interface VulnerabilityTagsComponentProps : Props {
     /**
      * Current logged-in user
      */
-    var userInfo: UserInfo?
+    var currentUserInfo: UserInfo?
 
     /**
      * flag to change
