+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/vuln/VulnerabilityView.kt
@@ -110,9 +110,7 @@ val vulnerabilityView: FC<VulnerabilitiesViewProps> = FC { props ->
                     this.vulnerability = vulnerability
                     this.setVulnerability = setVulnerability
                     this.fetchVulnerability = fetchVulnerability
-                    this.canEditVulnerability =
-                            permissions.isOwner || permissions.isSuperAdmin || permissions.isParticipant
-                    this.userInfo = props.currentUserInfo
+                    this.currentUserInfo = props.currentUserInfo
                     this.setIsEditDisabled = setIsEditDisabled
                     this.isEditDisabled = isEditDisabled
                     this.enrollRequest = enrollRequest
@@ -229,13 +227,3 @@ fun borderEditStyle(pointerCursor: Boolean = false): CSSProperties = jso {
         cursor = "pointer".unsafeCast<Cursor>()
     }
 }
-
-/**
- * @param currentUserInfo
- * @param vulnerability
- * @return if user can press edit button
- */
-internal fun hasRightsToEdit(currentUserInfo: UserInfo?, vulnerability: VulnerabilityExt) =
-        // globalRole is enriched on the App.kt level with "users/global-role" (based on ROLE from spring security authority)
-        // and is not stored in DB, so we cannot check userInfo directly, because in the getAllParticipants() users will have this field null
-        currentUserInfo.isSuperAdmin() || currentUserInfo?.name in vulnerability.getAllParticipants().map { it.name }
